+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - fix: Crash when converting Recrash Report #627
+- feat: Add SdkInfo to Envelope Header #626
 - fix: Deserialize envelope with header and item #620
 - fix: Set LogLevel in startWithConfigureOptions #613